requirements.txt
pandas
numpy
scikit-learn
dvc
mlflow
pyyaml
joblib
matplotlib
seaborn
